chore: fix up ipns example
