+++ packages/ipfs-core-config/package.json
@@ -95,7 +95,7 @@
     "it-drain": "^1.0.3",
     "libp2p-floodsub": "^0.27.0",
     "libp2p-gossipsub": "^0.11.1",
-    "libp2p-kad-dht": "^0.26.2",
+    "libp2p-kad-dht": "^0.26.4",
     "libp2p-mdns": "^0.17.0",
     "libp2p-mplex": "^0.10.2",
     "libp2p-tcp": "^0.17.1",
